Moved IStorage
